created directory for admin GUI, Controller and DAO; moved all function of admin to correct .dart file
