+++ src/lib/user_management/login_utente_gui.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 
 import 'package:civiconnect/theme.dart';
-import 'package:civiconnect/user_management/admin_main_gui.dart';
+import 'package:civiconnect/gestione_admin/admin_gui.dart';
 import 'package:civiconnect/user_management/registrazione_utente_gui.dart';
 import 'package:civiconnect/user_management/user_management_controller.dart';
 import 'package:civiconnect/widgets/input_textfield_decoration.dart';
